# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  domain.Product:
    properties:
      depth:
        type: string
      description:
        type: string
      height:
        type: string
      icon:
        type: string
      id:
        type: integer
      imageSrc:
        type: string
      length:
        type: string
      name:
        type: string
      price:
        type: integer
      status:
        type: integer
      title:
        type: string
      videoSrc:
        type: string
      weight:
        type: number
    type: object
...
  domain.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  domain.Support:
    properties:
      answerDescription:
        type: string
      id:
        type: integer
      problemDescription:
        type: string
      userID:
        type: integer
    type: object
  domain.UserInfo:
    properties:
      activity:
        type: integer
      birth:
        type: string
      email:
        type: string
      gender:
        type: integer
      height:
        type: integer
      id:
        type: integer
      name:
        type: string
      weight:
        type: integer
    type: object
  user.UserBasic:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /Supports/{userid}:
    get:
      description: get Support by userid
      operationId: get-support
      parameters:
      - description: user ID
        in: path
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Support'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetSupport
      tags:
      - support
...
  /products:
    get:
      description: get all products
      operationId: get-products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetProducts
      tags:
      - product
    post:
      consumes:
      - application/json
      description: create product
      operationId: create-product
      parameters:
      - description: Create product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateProduct
      tags:
      - product
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete product
      operationId: Delete-product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteProduct
      tags:
      - product
    get:
      description: get product by id
      operationId: get-product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetProduct
      tags:
      - product
    post:
      consumes:
      - application/json
      description: update product
      operationId: update-product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: update product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateProduct
      tags:
      - product
...
  /supports:
    get:
      description: get all Supports
      operationId: get-dupports
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Support'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetSupports
      tags:
      - support
    post:
      consumes:
      - application/json
      description: create support
      operationId: create-support
      parameters:
      - description: Create Support
        in: body
        name: Support
        required: true
        schema:
          $ref: '#/definitions/domain.Support'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Support'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: CreateSupport
      tags:
      - support
  /supports/{userid}:
    post:
      consumes:
      - application/json
      description: answer support
      operationId: answer-support
      parameters:
      - description: user ID
        in: path
        name: userid
        required: true
        type: integer
      - description: answer Support
        in: body
        name: Support
        required: true
        schema:
          $ref: '#/definitions/domain.Support'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Support'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: AnswerToSupport
      tags:
      - support
  /user:
    delete:
      consumes:
      - application/json
      description: Delete User
      operationId: Delete-User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteUser
      tags:
      - user
    get:
      description: get user
      operationId: get-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserBasic'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetUser
      tags:
      - user
  /user/information:
    get:
      description: get user info
      operationId: get-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: GetUserInfo
      tags:
      - user
    post:
      consumes:
      - application/json
      description: update user
      operationId: update-user
      parameters:
      - description: update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateUser
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"